refactor(realm): extract browser.fetch serialization from js-realm-shared
